restructured the files
